remove options.py
